blobs/get_blob: error on status != OK and display the HTTP body on client_error
The introduced code returns an error if the status code in the success
range. Before this change even error codes would be treated as success
and the HTTP body, likely an error description from the server, would be
misused as an image blob.

In accordance with [RFC2068][1] the body should be displayed to the user
on client errors. For usability we include a UTF8 decoded version since
it's likely to yield something human readable, e.g.:

```rust
ErrorMessage { msg: "could not download blob: GET request failed with
status \'404 Not Found\' and body of size 233: \"<!DOCTYPE HTML PUBLIC
\\\"-//W3C//DTD HTML 3.2 Final//EN\\\">\\n<title>404 Not
Found</title>\\n<h1>Not Found</h1>\\n<p>The requested URL was not found
on the server.  If you entered the URL manually please check your
spelling and try again.</p>\\n\"" }'
```

[1]: https://www.freesoft.org/CIE/RFC/2068/89.htm