v2: introduce a generic Manifest enum and impl
This type serves as an abstraction for the various manifest types and
makes it easier for consumers to retrieve information via a uniform
interface represented by the Manifest methods.

The motivation was the addition of a uniform way of retrieving the
architecture information of a given manifest.

This commit introduces this implementation for the ManifestSchema2*
type, which is now split into a spec compliant type and a container for
the compliant type and the config blob.