fix: support 'charset=utf-8' parameter for application/json mimetype
Seeing the charset param for application/json content-type present in the wild,
which was causing a strum error to be thrown when trying to construct the
MediaTypes enum from the header.

Following the existing solution above in the enum this simply adds a second
'serialize' attribute param to the ApplicationJson variant.

It is noted that steveeJ added a 'TODO' to find a generic way to handle these
on the existing usage, but this is not addressed in this bugfix.