add support for ManifestList images
ManifestList images also known as fat-manifests are
collection of manifests for different arch/os.
this pr adds support for ManifestList images.

As manifest list images only contain digests of the
images contained in the manifest, the `layers_digests`
function returns the digest of all the images
contained in the ManifestList instead of individual
layers of the manifests.
The layers of a specific image from manifest list can
be obtained using the digest of the image from the
manifest list and getting its manifest and manifestref
(get_manifest_and_ref()) and using this manifest of
the individual image to get the layers.