chore: Add struct to handle both access_token and token from JSON
In contrast to ACR, Docker Hub responds with `token` and `access_token` which is not supported by the alias attribute provided by serde. That is why I came up with a dedicated struct deserializing both fields into an Option<String> the TryFrom<> implementation first tries to use the token property (as it was before this PR). If that is not present, it tries to take the access_token property. If both are missing, a corresponding error is thrown

Signed-off-by: Thorsten Hans <thorsten.hans@fermyon.com>